+++ NEWS.md
@@ -1,3 +1,5 @@
+# healthyR.ai (development version)
+
 # healthyR.ai 0.0.7
 
 ## Breaking Changes