+++ NEWS.md
@@ -9,6 +9,8 @@ None
 2. Fix #227 - Add function `hai_auto_wflw_metrics()`
 3. Fix #243 - Add boilerplate functions for C5.0 `hai_c50_data_prepper()` and
 `hai_auto_c50()`
+4. Fix #246 - Add boilerplate functions for glmnet `hai_glmnet_data_prepper()` and
+`hai_auto_glmnet()`
 
 ## Minor Fixes and Improvements
 1. Fix #240 = Add parameter arguments as attributes to list output for boilerplate