+++ NEWS.md
@@ -15,6 +15,8 @@ None
 `hai_auto_cubist()`
 6. Fix #256 - Add boilerplate functions for Earth `hai_earth_data_prepper()` and
 `hai_auto_earth()`
+7. Fix #259 - Add boilerplate functions for Kernlab `hai_svm_poly_data_prepper()` and
+`hai_auto_svm_poly()`
 
 ## Minor Fixes and Improvements
 1. Fix #240 = Add parameter arguments as attributes to list output for boilerplate