+++ NEWS.md
@@ -21,8 +21,9 @@ None
 `hai_auto_svm_rbf()`
 
 ## Minor Fixes and Improvements
-1. Fix #240 = Add parameter arguments as attributes to list output for boilerplate
+1. Fix #240 - Add parameter arguments as attributes to list output for boilerplate
 functions.
+2. Fix #273 - We fixed a bug 🐛 in step_hyperbolic() in tidymodels/recipes#932 and this PR updates healthyR.ai to reflect that fix. Thank you @juliasilge
 
 # healthyR.ai 0.0.7